+++ src/test/kotlin/org/jetbrains/intellij/IntelliJPluginSpec.kt
@@ -16,7 +16,6 @@ import org.jetbrains.intellij.IntelliJPluginConstants.JAR_SEARCHABLE_OPTIONS_TAS
 import org.jetbrains.intellij.IntelliJPluginConstants.LIST_BUNDLED_PLUGINS_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.LIST_PRODUCTS_RELEASES_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.MARKETPLACE_HOST
-import org.jetbrains.intellij.IntelliJPluginConstants.MINIMAL_SUPPORTED_GRADLE_VERSION
 import org.jetbrains.intellij.IntelliJPluginConstants.PATCH_PLUGIN_XML_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.PLUGIN_GROUP_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.POST_INSTRUMENT_CODE_TASK_NAME
@@ -576,12 +575,12 @@ class IntelliJPluginSpec : IntelliJPluginSpecBase() {
     }
 
     @Test
-    fun `expect successful build using minimal supported Gradle version`() {
-        val buildResult = build(MINIMAL_SUPPORTED_GRADLE_VERSION, false, "help")
-
-        assertContains("BUILD SUCCESSFUL", buildResult.output)
-        assertNotContains("Gradle IntelliJ Plugin requires Gradle", buildResult.output)
-    }
+//    fun `expect successful build using minimal supported Gradle version`() {
+//        val buildResult = build(MINIMAL_SUPPORTED_GRADLE_VERSION, false, "help")
+//
+//        assertContains("BUILD SUCCESSFUL", buildResult.output)
+//        assertNotContains("Gradle IntelliJ Plugin requires Gradle", buildResult.output)
+//    }
 
     @SuppressWarnings("GrEqualsBetweenInconvertibleTypes")
     fun assertPathParameters(testCommand: ProcessProperties, sandboxPath: String) {