+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `custom(String)` helper of `intellij.pluginRepositories` configuration requires now passing a direct URL to the `updatePlugins.xml` file [#1252](../../issues/1252)
 - `listProductsReleases` task doesn't print output anymore
 - `listBundledPlugins` task doesn't print output anymore
+- Set minimum supported Gradle version from `7.1` to `7.3`
 
 ### Fixed
 - Replace `Contents/Contents` part within JVM arguments to a single `Contents` – happens with macOS distribution